requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==4.2.16
djangorestframework==3.15.2
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
multidict==6.1.0
openai==0.28.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
